feature: better latitude build command and some other shenanigans
Improve overall code of Latitude CLI. Added prepare step in order to
decouple docker from our build process.

+++ packages/cli/src/commands/build/index.ts
@@ -1,29 +1,37 @@
 import colors from 'picocolors'
 import config from '../../config'
 import path from 'path'
-import syncQueries from '../shared/syncQueries'
-import syncViews from '../shared/syncViews'
 import { APP_FOLDER } from '../constants'
 import { onError } from '../../utils'
 import { spawn } from 'child_process'
 import maybeSetupApp from '../shared/maybeSetupApp'
 import fs from 'fs-extra'
+import prepareCommand from '../prepare'
 
-export default async function build() {
+export default async function build(
+  { docker = false }: { docker?: boolean } = { docker: false },
+) {
   const ready = await maybeSetupApp()
   if (!ready) process.exit(1)
 
   const cwd = config.cwd
-  const appName = path.basename(cwd)
 
-  await syncViews({ dataAppDir: cwd, appName })
-  await syncQueries({ rootDir: cwd })
+  await prepareCommand()
 
   const buildCwd = path.join(cwd, APP_FOLDER)
-  const buildProcess = spawn(config.pkgManager.command, ['run', 'build'], {
-    detached: false,
-    cwd: buildCwd,
-  })
+  let buildProcess
+
+  if (docker) {
+    buildProcess = spawn('docker', ['build', '.'], {
+      detached: false,
+      cwd,
+    })
+  } else {
+    buildProcess = spawn(config.pkgManager.command, ['run', 'build'], {
+      detached: false,
+      cwd: buildCwd,
+    })
+  }
 
   buildProcess.stdout?.on('data', (data) => {
     console.log(colors.gray(data))
@@ -48,32 +56,37 @@ export default async function build() {
       process.exit(code)
     }
 
-    // symlink the .latitude/app folder to a ./build folder
-    // so that we can deploy the build folder
-    const targetPath = path.join(cwd, 'build')
-    const appPath = path.join(cwd, '.latitude', 'app')
+    if (!docker) {
+      // symlink the .latitude/app folder to a ./build folder
+      // so that we can deploy the build folder
+      const targetPath = path.join(cwd, 'build')
+      const appPath = path.join(cwd, '.latitude', 'app')
 
-    try {
-      const exists = await fs.pathExists(targetPath)
-      if (exists) await fs.remove(targetPath)
+      try {
+        const exists = await fs.pathExists(targetPath)
+        if (exists) await fs.remove(targetPath)
 
-      await fs.symlink(appPath, targetPath)
-    } catch (e) {
-      console.log('error: ', e)
-      onError({
-        error: e as Error,
-        message: 'Error creating symlink to build folder',
-      })
+        await fs.symlink(appPath, targetPath)
+      } catch (e) {
+        console.log('error: ', e)
+        onError({
+          error: e as Error,
+          message: 'Error creating symlink to build folder',
+        })
+      }
     }
 
     console.log(colors.green(`📦 Build completed!\n`))
-    console.log(
-      colors.gray(`
-      To run your app, use the following commands:\n
-      $ cd build
-      $ node build
-    `),
-    )
+
+    if (!docker) {
+      console.log(
+        colors.gray(`
+        To run your app, use the following commands:\n
+        $ cd build
+        $ node build
+      `),
+      )
+    }
 
     process.exit()
   })
